[CBW-910] Update terms & conditions screen (#310)
[CBW-910] handle latest t&c request

Until now each of the wallet types have had their own T&C’s, and in some of the wallets they have been hardcoded making it troublesome to make updates to the T&Cs. Instead, we decided to make a common T&C that is shared between the browser wallet and the recovery phrase based wallets for iOS and Android.