WalletConnect account selection view (#328)
Added initial integration with WalletConnect Sign client: After scanning
the QR code it shows the account list, which brings you to the
approve/reject screen. If approved, the connection is established.
Requests from the client side are not yet handled.

This is a snapshot of an implementation that isn't yet done. Disconnects
don't work properly and you easily get into a state with duplicate
events firing, causing an erratic experience.

---------

Co-authored-by: Michael Bisgaard Olesen <mo@concordium.com>