[CBW-1218] Block WalletConnect multiple requests at the same time (#382)
## Purpose
For safety, it is important to not allow dApps to send requests upon
other request. It can mislead a user and make him accept request, that
he/she would've otherwise not approved.

## Changes
In order to prevent that, a bool flag has been added and is set `true`
when a request is pending. After it's rejected or accepted, the flag is
set back to false.

Jira ticket: https://concordium.atlassian.net/browse/CBW-1218

---------

Co-authored-by: Michael Bisgaard Olesen <mo@concordium.com>