[CBW-1344] Implement sorting for cached tokens in account details screen. (#383)
## Purpose
Tokens open in different order every time Fungible/Collectibles pages
are open and are not sorted at all.
## Changes
I added `dateAdded` property to `CIS2TokenOwnershipEntity` that
represents the date of adding the token to database.
I implemented sorting function that sorts token elements by adding date
to database, from oldest to newest. There might be some elements that
have been added to database prior to this. In that case they're sorted
alphabetically and displayed below the ones with date.
## Task
https://concordium.atlassian.net/browse/CBW-1344