Fix updating pool settings while update validator stake (#396)
## Purpose
Commission rates are set to 100% in case they are not explicitly set in
the validators transaction. Prevent changing commission rates while
updates validator stake

## Changes

Add ```private func loadPoolParameters(for bakerID: Int) {...}``` method
to retrieve validator pool commissionRates in case if profile is
`validator`, or load them from chain parameters like it was before. So
in this case there will not be changes rates to 100%, while user change
only validator stake.

---------

Co-authored-by: Michael Bisgaard Olesen <mo@concordium.com>